requirements/prod.txt
# Everything needed in production

# Flask
Flask==0.10.1
MarkupSafe==0.23
Werkzeug==0.10.4
Jinja2==2.7.3
itsdangerous==0.24
factory-boy==2.5.1

# Database
Flask-SQLAlchemy==2.0
SQLAlchemy==0.9.8
psycopg2==2.6.1

# Migrations
Flask-Migrate==1.3.1

# Forms
Flask-WTF==0.11
WTForms==2.0.2

# Deployment
gunicorn>=19.1.1
flask-sslify>=0.1.5

# Assets
Flask-Assets==0.10
cssmin>=0.2.0
jsmin>=2.0.11

# Auth
Flask-Login==0.2.11
Flask-Bcrypt==0.6.2

# Caching
Flask-Cache>=0.13.1

# Debug toolbar
Flask-DebugToolbar==0.9.2
